Refactor the Rosie disable comment fix. Move tests to a dedicated class.
